backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
scikit-learn
Pillow
tqdm
pandas
opencv-python